+++ packages/web/design-system/src/ui/config-provider/index.ts
@@ -1,2 +1,3 @@
 export { useInjectConfigContext } from '@/ui/config-provider/config.context'
+export type { ConfigProviderProps } from '@/ui/config-provider/config.types'
 export { default as UIConfigProvider } from '@/ui/config-provider/ConfigProvider.vue'