+++ packages/web/design-system/src/ui/field-wrapper/index.ts
@@ -1,2 +1,2 @@
-export { type FieldWrapperProps as UIFieldWrapperProps } from './fieldWrapper.props.ts'
+export { type FieldWrapperProps as UIFieldWrapperProps } from './fieldWrapper.props'
 export { default as UIFieldWrapper } from './FieldWrapper.vue'