+++ packages/web/design-system/src/ui/page/detailPane.type.ts
@@ -1,5 +1,7 @@
 export type DetailPaneStorageStrategy = 'localStorage' | 'routeQuery'
 
+export type DetailPaneVariant = | 'bordered-inline' | 'bordered-overlay' | 'full-height-inline' | 'full-height-overlay'
+
 export interface DetailPaneStorage {
   key: string
   strategy: DetailPaneStorageStrategy
@@ -8,4 +10,5 @@ export interface DetailPaneStorage {
 export interface DetailPaneConfig {
   isResizable?: boolean
   storage?: DetailPaneStorage | null
+  variant?: DetailPaneVariant
 }