+++ packages/web/design-system/src/ui/textarea-field/index.ts
@@ -1,2 +1,2 @@
-export { type TextareaFieldProps as UITextareaFieldProps } from './textareaField.props.ts'
+export { type TextareaFieldProps as UITextareaFieldProps } from './textareaField.props'
 export { default as UITextareaField } from './TextareaField.vue'